+++ package.json
@@ -13,7 +13,7 @@
     }
   },
   "scripts": {
-    "test": "npm run test"
+    "test": "vitest"
   },
   "repository": {
     "type": "git",
@@ -27,5 +27,10 @@
   "bugs": {
     "url": "https://github.com/oakstudios/inline-svg/issues"
   },
-  "homepage": "https://github.com/oakstudios/inline-svg#readme"
+  "homepage": "https://github.com/oakstudios/inline-svg#readme",
+  "devDependencies": {
+    "@vitest/ui": "^0.29.8",
+    "vite": "^4.2.1",
+    "vitest": "^0.29.8"
+  }
 }
