+++ vite.config.js
@@ -1,0 +1,12 @@
+/// <reference types="vitest" />
+
+// Configure Vitest (https://vitest.dev/config/)
+
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  test: {
+    /* for example, use global to avoid globals imports (describe, test, expect): */
+    // globals: true,
+  },
+})
